update validate and response register, login

+++ src/main/java/com/be_uterace/payload/response/ClubRankingResponse.java
@@ -13,6 +13,6 @@ public class ClubRankingResponse {
     private String name;
     private String image;
     private Double total_distance;
-    private Long total_members;
+    private Integer total_members;
     private Long total_activities;
 }
